% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmm.R
\name{extract_donor_id}
\alias{extract_donor_id}
\title{Extract recorded PMM donor IDs}
\usage{
extract_donor_id(data, variable)
}
\arguments{
\item{data}{A \code{mids} object containing a \code{pmmrw} imputation.}

\item{variable}{Name of the PMM-imputed variable.}
}
\value{
An \code{n} by \code{m} matrix of donor row IDs.
}
\description{
Extract recorded PMM donor IDs
}
