% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{giganti_data}
\alias{giganti_data}
\title{Giganti and Shepherd simulation data}
\format{
A data frame with 4,000 rows and variables \code{ID}, \code{X1}, \code{X2},
\code{A.star}, \code{D.star}, \code{A}, \code{D}, and \code{Intercept}.
}
\usage{
giganti_data
}
\description{
A simulated two-phase dataset based on Giganti and Shepherd (2020).
}
\keyword{datasets}
