% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_rw.R
\name{with_rw}
\alias{with_rw}
\title{Fit an analysis model to multiply imputed data}
\usage{
with_rw(data, expr)
}
\arguments{
\item{data}{A \code{mids} object created with \code{mice(..., tasks = "train")} using
\code{norm}, \code{logreg}, or \code{pmmrw} imputation.}

\item{expr}{An \code{lm} or binomial \code{glm} expression.}
}
\value{
An \code{rw_fit} object containing the fitted models and RW components.
}
\description{
Fit an analysis model to multiply imputed data
}
