% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_rw.R
\name{pool_rw}
\alias{pool_rw}
\title{Pool an RW analysis}
\usage{
pool_rw(object)
}
\arguments{
\item{object}{An \code{rw_fit} object returned by \code{\link[=with_rw]{with_rw()}}.}
}
\value{
An \code{rw_pool} object containing pooled estimates and RW variance.
}
\description{
Pool an RW analysis
}
